refactor getUserInfo api and adjust the types
