+++ src/container/Chat/components/ChatroomSection/ChatroomSection.tsx
@@ -6,14 +6,12 @@ import { useGetUserInfo } from '~/api/user/user'
 import { Props } from './types'
 import { useInView } from 'react-intersection-observer'
 
-const ChatroomSection = ({ chatroomId }: Props) => {
+const ChatroomSection = ({ chatroomId, userData }: Props) => {
     const {
         data: chatroomListData,
         hasNextPage,
         fetchNextPage,
     } = useGetChatroomList()
-
-    const { data: userData } = useGetUserInfo()
 
     const { ref, inView } = useInView({
         threshold: 0.5,
@@ -42,12 +40,12 @@ const ChatroomSection = ({ chatroomId }: Props) => {
                         key={data.id}
                         id={data.id}
                         userName={
-                            userData?.user?.id === data.mentor.id
+                            userData?.id === data.mentor.id
                                 ? data.member.userName
                                 : data.mentor.userName
                         }
                         avatar={
-                            userData?.user?.id === data.mentor.id
+                            userData?.id === data.mentor.id
                                 ? data.member.avatar
                                 : data.mentor.avatar
                         }
