+++ src/container/Chat/components/ChattingArea/components/types.ts
@@ -1,0 +1,4 @@
+export interface Props {
+    avatar: string
+    userName: string
+}
