+++ src/container/Chat/components/ChattingArea/types.ts
@@ -1,7 +1,9 @@
 import { MessageData } from '~/lib/types'
+import { UserForMentor, UserForMember } from '~/api/user/types'
 
 export interface Props {
     chatroomId: string
+    userData: UserForMember | UserForMentor
 }
 
 export interface QueryParams {
