+++ src/container/UserProfile/MemberPage/types.ts
@@ -1,23 +1,5 @@
-import { GenderCode } from '../types'
+import { UserForMember } from '~/api/user/types'
 
 export interface Props {
-    userData: MemberProfileData
-}
-
-export interface MemberProfileData {
-    userName: string
-    email: string
-    avatar: string
-    gender: GenderCode
-    country: string
-    title: string
-    company: string
-    phoneNumber: string
-    introduction: string
-    level: number
-    fieldOfWork: string[]
-    id: string
-    emailOtp: boolean
-    created_at: string
-    updated_at: string
+    userData: UserForMember
 }
