+++ src/container/UserProfile/UserProfile.tsx
@@ -1,7 +1,24 @@
-import UserProfileSection from './UserProfileSection/UserProfileSection'
+import { callGetUserInfoHandler } from '~/api/user/userAPI'
+import MentorPage from './MentorPage/MentorPage'
+import MemberPage from './MemberPage/MemberPage'
+import { USER_IDENTITY } from '~/api/user/types'
 
-const UserProfile = () => {
-    return <UserProfileSection />
+const UserProfile = async () => {
+    const userData = await callGetUserInfoHandler()
+
+    return (
+        <>
+            {userData && (
+                <>
+                    {userData.identity === USER_IDENTITY.MENTOR ? (
+                        <MentorPage userData={userData.user} />
+                    ) : (
+                        <MemberPage userData={userData.user} />
+                    )}
+                </>
+            )}
+        </>
+    )
 }
 
 export default UserProfile
