+++ src/container/VideoConference/components/VideoConferenceSection/types.ts
@@ -1,6 +1,6 @@
 import { UserForMentor, UserForMember } from '~/api/user/types'
 
 export interface Props {
-    chatroomId: string
-    userData: UserForMember | UserForMentor
+    roomId: string
+    userData: UserForMentor | UserForMember
 }
